use kind argument directly and rename function
