+++ hooli_batch_enrichment/dagster_batch_enrichment/assets.py
@@ -17,10 +17,7 @@ class experimentConfig(Config):
     )
 
 @asset(
-    tags={
-        **build_kind_tag("Kubernetes"),
-        **build_kind_tag("S3"),
-        },
+    kinds={"Kubernetes", "S3"}
 )
 def raw_data(
     context: OpExecutionContext, 
@@ -93,10 +90,7 @@ def concat_chunk_list(chunks) -> pd.DataFrame:
 
 
 @graph_asset(
-    tags={
-        **build_kind_tag("Kubernetes"),
-        **build_kind_tag("S3"),
-        },
+    kinds={"Kubernetes", "S3"}
 )
 def enriched_data(raw_data) -> pd.DataFrame:
     """Full enrichment process"""
